requirements.txt
Django
Faker
django-debug-toolbar
django-zen-queries
